requirements.txt
Django==3.1.4
PyMySQL==0.10.1
djangorestframework==3.12.4
